requirements.txt
python-dotenv>=0.21.0
requests>=2.28.0
openai>=1.0.0
python-pptx>=0.6.21
pytest>=7.0.0
tqdm>=4.0.0
serpapi>=0.1.5
google-generativeai>=0.8.0
Pillow>=10.0.0